Add app name, moved mapComponent into components
